career-ai/backend/app/db/supabase.py
import os
from dotenv import load_dotenv
from supabase import create_client

# Load environment variables
load_dotenv()

# Initialize Supabase client
def get_supabase_client():
    """Get a Supabase client instance"""
    supabase_url = os.environ.get("SUPABASE_URL")
    supabase_key = os.environ.get("SUPABASE_KEY")
    
    if not supabase_url or not supabase_key:
        raise ValueError("Supabase URL and key must be set as environment variables")
    
    return create_client(supabase_url, supabase_key)
