career-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==1.0.4
groq==0.4.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
